Rename {cap-release,cap-pre-release}.yaml.tmpl to .yaml
This allows for linting and easy editor highlighting, etc, as there's no need to
have them as templates; they are config files afterall.

Load the config context in the "config" variable containing the context from the
"config" datasource. It would be possible to load the config as a '.' context,
but the context is lost when in "range" or "with", and would need to be
addressed with $.key, which seems less clear than $config.key as we have with
the datasource.