Provide no-op on deploy-k8s job when caasp4, for now
CaaSP4 clusters are currently pets: we create them out of the pipeline, be it in
ECP (we need to build a skuba image, and configure dind for concourse) or be it
KVM (not added to catapult yet).

Hence, for caasp4 clusters, we need to not run deploy-k8s job.

Sadly, the pipeline definitions don't have a simple way to disable one job for 1
backend only.

We could deploy additional
  {cap-pre-release,cap-release,cap-release-upgrades,…}-caasp4.yaml
pipelines, but it's messy. Or we could change the gomplate templating, for which we
don't have time now.

Hence, make deploy-k8s do nothing when caasp4, for now.