Add {cap*release*}-caasp.yaml pipelines for now
Add copies of all pipelines, for caasp.

These pipelines have the deploy-k8s job disabled, as the clusters are pets.

We are in a pickle:
- We cannot deploy cap pipelines where 1 provider has 1 job
  less because of the structure of the template.
- we cannot add a no-op on deploy-k8s when caasp because of the `put` step on
  it (see https://github.com/SUSE/cap/pull/60).

We could redo the templating right now (no), or we can just create copies of the
pipelines, configured with only caasp4 and without the deploy-k8s job. Which is
what this commit does.